relocates setCheckDetails function to checks file
